COMMON: Split utility and ec files.
Both, utility.c and ec_curves.c required a lot of additional PKCS #11
definitions which made it hard to use them in several places.  Split the files
such that independent parts can be reused without requiring additional PKCS
