+++ usr/lib/ica_s390_stdll/ica_s390_stdll.mk
@@ -35,7 +35,8 @@ opencryptoki_stdll_libpkcs11_ica_la_SOURCES =				\
 	usr/lib/common/mech_list.c usr/lib/common/shared_memory.c	\
 	usr/lib/common/profile_obj.c usr/lib/common/attributes.c	\
 	usr/lib/ica_s390_stdll/ica_specific.c usr/lib/common/dlist.c	\
-	usr/lib/common/mech_openssl.c
+	usr/lib/common/mech_openssl.c					\
+	usr/lib/common/utility_common.c usr/lib/common/ec_supported.c
 
 if ENABLE_LOCKS
 opencryptoki_stdll_libpkcs11_ica_la_SOURCES +=				\
