+++ usr/lib/icsf_stdll/icsf_stdll.mk
@@ -25,30 +25,31 @@ opencryptoki_stdll_libpkcs11_icsf_la_LDFLAGS =				\
 	-lrt -llber							\
 	-Wl,--version-script=${srcdir}/opencryptoki_tok.map
 
-opencryptoki_stdll_libpkcs11_icsf_la_SOURCES =				\
-	usr/lib/common/asn1.c usr/lib/common/dig_mgr.c			\
-	usr/lib/common/hwf_obj.c usr/lib/common/trace.c			\
-	usr/lib/common/key.c usr/lib/common/mech_dh.c			\
-	usr/lib/common/mech_rng.c usr/lib/common/sign_mgr.c		\
-	usr/lib/common/cert.c usr/lib/common/dp_obj.c			\
-	usr/lib/common/mech_aes.c usr/lib/common/mech_rsa.c		\
-	usr/lib/common/mech_ec.c usr/lib/common/obj_mgr.c		\
-	usr/lib/common/template.c usr/lib/common/p11util.c		\
-	usr/lib/common/data_obj.c usr/lib/common/encr_mgr.c		\
-	usr/lib/common/key_mgr.c usr/lib/common/mech_md2.c		\
-	usr/lib/common/mech_sha.c usr/lib/common/object.c		\
-	usr/lib/common/decr_mgr.c usr/lib/common/globals.c		\
-	usr/lib/common/sw_crypt.c usr/lib/common/loadsave.c		\
-	usr/lib/common/utility.c usr/lib/common/mech_des.c		\
-	usr/lib/common/mech_des3.c usr/lib/common/mech_md5.c		\
-	usr/lib/common/mech_ssl3.c usr/lib/common/verify_mgr.c		\
-	usr/lib/common/mech_list.c usr/lib/common/shared_memory.c	\
-	usr/lib/common/attributes.c usr/lib/icsf_stdll/new_host.c	\
-	usr/lib/common/profile_obj.c usr/lib/common/dlist.c		\
-	usr/lib/icsf_stdll/pbkdf.c usr/lib/icsf_stdll/icsf_specific.c	\
+opencryptoki_stdll_libpkcs11_icsf_la_SOURCES = usr/lib/common/asn1.c	\
+	usr/lib/common/dig_mgr.c usr/lib/common/hwf_obj.c		\
+	usr/lib/common/trace.c usr/lib/common/key.c			\
+	usr/lib/common/mech_dh.c usr/lib/common/mech_rng.c		\
+	usr/lib/common/sign_mgr.c usr/lib/common/cert.c			\
+	usr/lib/common/dp_obj.c usr/lib/common/mech_aes.c		\
+	usr/lib/common/mech_rsa.c usr/lib/common/mech_ec.c		\
+	usr/lib/common/obj_mgr.c usr/lib/common/template.c		\
+	usr/lib/common/p11util.c usr/lib/common/data_obj.c		\
+	usr/lib/common/encr_mgr.c usr/lib/common/key_mgr.c		\
+	usr/lib/common/mech_md2.c usr/lib/common/mech_sha.c		\
+	usr/lib/common/object.c usr/lib/common/decr_mgr.c		\
+	usr/lib/common/globals.c usr/lib/common/sw_crypt.c		\
+	usr/lib/common/loadsave.c usr/lib/common/utility.c		\
+	usr/lib/common/mech_des.c usr/lib/common/mech_des3.c		\
+	usr/lib/common/mech_md5.c usr/lib/common/mech_ssl3.c		\
+	usr/lib/common/verify_mgr.c usr/lib/common/mech_list.c		\
+	usr/lib/common/shared_memory.c usr/lib/common/attributes.c	\
+	usr/lib/icsf_stdll/new_host.c usr/lib/common/profile_obj.c	\
+	usr/lib/common/dlist.c usr/lib/icsf_stdll/pbkdf.c		\
+	usr/lib/icsf_stdll/icsf_specific.c				\
 	usr/lib/icsf_stdll/icsf_config_parse.y				\
 	usr/lib/icsf_stdll/icsf_config_lexer.l				\
-	usr/lib/icsf_stdll/icsf.c
+	usr/lib/icsf_stdll/icsf.c usr/lib/common/utility_common.c	\
+	usr/lib/common/ec_supported.c
 
 if ENABLE_LOCKS
 opencryptoki_stdll_libpkcs11_icsf_la_SOURCES +=				\
