+++ usr/lib/tpm_stdll/tpm_stdll.mk
@@ -35,7 +35,8 @@ opencryptoki_stdll_libpkcs11_tpm_la_SOURCES =				\
 	usr/lib/common/shared_memory.c usr/lib/common/profile_obj.c	\
 	usr/lib/tpm_stdll/tpm_specific.c usr/lib/common/attributes.c	\
 	usr/lib/tpm_stdll/tpm_openssl.c usr/lib/tpm_stdll/tpm_util.c	\
-	usr/lib/common/dlist.c usr/lib/common/mech_openssl.c
+	usr/lib/common/dlist.c usr/lib/common/mech_openssl.c		\
+	usr/lib/common/utility_common.c usr/lib/common/ec_supported.c
 
 if ENABLE_LOCKS
 opencryptoki_stdll_libpkcs11_tpm_la_SOURCES +=				\
