+++ usr/sbin/pkcscca/pkcscca.mk
@@ -17,29 +17,28 @@ usr_sbin_pkcscca_pkcscca_CFLAGS  =					\
 	-I${srcdir}/usr/lib/common					\
 	-I${srcdir}/usr/sbin/pkcscca
 
-usr_sbin_pkcscca_pkcscca_SOURCES =					\
-	usr/lib/common/asn1.c usr/lib/common/dig_mgr.c			\
-	usr/lib/common/hwf_obj.c usr/lib/common/trace.c			\
-	usr/lib/common/key.c usr/lib/common/mech_list.c			\
-	usr/lib/common/mech_dh.c usr/lib/common/sign_mgr.c		\
-	usr/lib/common/cert.c usr/lib/common/dp_obj.c			\
-	usr/lib/common/mech_aes.c usr/lib/common/mech_rsa.c		\
-	usr/lib/common/mech_ec.c usr/lib/common/obj_mgr.c		\
-	usr/lib/common/template.c usr/lib/common/data_obj.c		\
-	usr/lib/common/encr_mgr.c usr/lib/common/key_mgr.c		\
-	usr/lib/common/mech_md2.c usr/lib/common/mech_sha.c		\
-	usr/lib/common/object.c	usr/lib/common/decr_mgr.c		\
-	usr/lib/common/globals.c usr/lib/common/loadsave.c		\
-	usr/lib/common/utility.c usr/lib/common/mech_des.c		\
-	usr/lib/common/mech_des3.c usr/lib/common/mech_md5.c		\
-	usr/lib/common/mech_ssl3.c usr/lib/common/verify_mgr.c		\
-	usr/lib/common/p11util.c usr/lib/common/sw_crypt.c		\
-	usr/lib/common/shared_memory.c usr/lib/common/profile_obj.c	\
-	usr/lib/common/attributes.c usr/lib/common/mech_rng.c		\
-	usr/lib/common/pkcs_utils.c  usr/lib/common/dlist.c		\
-	usr/sbin/pkcscca/pkcscca.c
-	
-	
+usr_sbin_pkcscca_pkcscca_SOURCES = usr/lib/common/asn1.c		\
+	usr/lib/common/dig_mgr.c usr/lib/common/hwf_obj.c		\
+	usr/lib/common/trace.c usr/lib/common/key.c			\
+	usr/lib/common/mech_list.c usr/lib/common/mech_dh.c		\
+	usr/lib/common/sign_mgr.c usr/lib/common/cert.c			\
+	usr/lib/common/dp_obj.c usr/lib/common/mech_aes.c		\
+	usr/lib/common/mech_rsa.c usr/lib/common/mech_ec.c		\
+	usr/lib/common/obj_mgr.c usr/lib/common/template.c		\
+	usr/lib/common/data_obj.c usr/lib/common/encr_mgr.c		\
+	usr/lib/common/key_mgr.c usr/lib/common/mech_md2.c		\
+	usr/lib/common/mech_sha.c usr/lib/common/object.c		\
+	usr/lib/common/decr_mgr.c usr/lib/common/globals.c		\
+	usr/lib/common/loadsave.c usr/lib/common/utility.c		\
+	usr/lib/common/mech_des.c usr/lib/common/mech_des3.c		\
+	usr/lib/common/mech_md5.c usr/lib/common/mech_ssl3.c		\
+	usr/lib/common/verify_mgr.c usr/lib/common/p11util.c		\
+	usr/lib/common/sw_crypt.c usr/lib/common/shared_memory.c	\
+	usr/lib/common/profile_obj.c usr/lib/common/attributes.c	\
+	usr/lib/common/mech_rng.c usr/lib/common/pkcs_utils.c		\
+	usr/lib/common/dlist.c usr/sbin/pkcscca/pkcscca.c		\
+	usr/lib/common/utility_common.c usr/lib/common/ec_supported.c
+
 if ENABLE_LOCKS
 usr_sbin_pkcscca_pkcscca_SOURCES +=				\
 	usr/lib/common/lock_btree.c usr/lib/common/lock_sess_mgr.c
